Corrected some issues related to the update period.
Differentiated between negative and zero values for update period.

Update period is now in milliseconds.

+++ src/net/xqhs/util/logging/output/ConsoleOutput.java
@@ -13,7 +13,7 @@ import net.xqhs.util.logging.Logger.Level;
 public class ConsoleOutput implements StreamLogOutput {
 	
 	@Override
-	public int getUpdatePeriod() {
+	public long getUpdatePeriod() {
 		return 0;
 	}
